+++ sitejuicer/__init__.py
@@ -2,7 +2,7 @@
 SiteJuicer - utility to convert URLs to markdown using Jina Reader
 """
 
-__version__ = '0.2.2'
+__version__ = '0.2.6'
 
 # Import key functions for easier access
 from sitejuicer.core import fetch_content, save_markdown